ott-admin/backend/player_registration/requirements.txt
aiohttp==3.8.1
fastapi==0.71.0
loguru==0.6.0
pydantic==1.9.0
uvicorn==0.16.0
pytest==7.0.1
boto3
kinesis-python
redis-py-cluster==2.1.3
requests==2.28.1
